+++ frontend/src/components/Explore.vue
