+++ frontend/src/components/Trending.vue
